renamed add new login fragment and vm as well
