+++ app/src/main/java/com/andryoga/safebox/data/repository/BankCardDataRepositoryImpl.kt
@@ -4,7 +4,7 @@ import com.andryoga.safebox.data.db.docs.SearchBankCardData
 import com.andryoga.safebox.data.db.entity.BankCardDataEntity
 import com.andryoga.safebox.data.db.secureDao.BankCardDataDaoSecure
 import com.andryoga.safebox.data.repository.interfaces.BankCardDataRepository
-import com.andryoga.safebox.ui.view.home.addNewData.bankCard.AddNewBankCardScreenData
+import com.andryoga.safebox.ui.view.home.dataDetails.bankCard.AddNewBankCardScreenData
 import kotlinx.coroutines.flow.Flow
 import java.util.*
 import javax.inject.Inject
