+++ app/src/main/java/com/andryoga/safebox/data/repository/SecureNoteDataRepositoryImpl.kt
@@ -4,7 +4,7 @@ import com.andryoga.safebox.data.db.docs.SearchSecureNoteData
 import com.andryoga.safebox.data.db.entity.SecureNoteDataEntity
 import com.andryoga.safebox.data.db.secureDao.SecureNoteDataDaoSecure
 import com.andryoga.safebox.data.repository.interfaces.SecureNoteDataRepository
-import com.andryoga.safebox.ui.view.home.addNewData.secureNote.SecureNoteScreenData
+import com.andryoga.safebox.ui.view.home.dataDetails.secureNote.SecureNoteScreenData
 import kotlinx.coroutines.flow.Flow
 import java.util.*
 import javax.inject.Inject
