+++ app/src/main/java/com/andryoga/safebox/data/repository/interfaces/BankCardDataRepository.kt
@@ -1,7 +1,7 @@
 package com.andryoga.safebox.data.repository.interfaces
 
 import com.andryoga.safebox.data.db.docs.SearchBankCardData
-import com.andryoga.safebox.ui.view.home.addNewData.bankCard.AddNewBankCardScreenData
+import com.andryoga.safebox.ui.view.home.dataDetails.bankCard.AddNewBankCardScreenData
 import kotlinx.coroutines.flow.Flow
 
 interface BankCardDataRepository {
