+++ app/src/main/java/com/andryoga/safebox/data/repository/interfaces/LoginDataRepository.kt
@@ -1,7 +1,7 @@
 package com.andryoga.safebox.data.repository.interfaces
 
 import com.andryoga.safebox.data.db.docs.SearchLoginData
-import com.andryoga.safebox.ui.view.home.addNewData.login.LoginScreenData
+import com.andryoga.safebox.ui.view.home.dataDetails.login.LoginScreenData
 import kotlinx.coroutines.flow.Flow
 
 interface LoginDataRepository {
