+++ app/src/main/java/com/andryoga/safebox/data/repository/interfaces/SecureNoteDataRepository.kt
@@ -1,7 +1,7 @@
 package com.andryoga.safebox.data.repository.interfaces
 
 import com.andryoga.safebox.data.db.docs.SearchSecureNoteData
-import com.andryoga.safebox.ui.view.home.addNewData.secureNote.SecureNoteScreenData
+import com.andryoga.safebox.ui.view.home.dataDetails.secureNote.SecureNoteScreenData
 import kotlinx.coroutines.flow.Flow
 
 interface SecureNoteDataRepository {
