+++ app/src/main/java/com/andryoga/safebox/ui/view/home/dataDetails/secureNote/SecureNoteScreenData.kt
@@ -1,4 +1,4 @@
-package com.andryoga.safebox.ui.view.home.addNewData.secureNote
+package com.andryoga.safebox.ui.view.home.dataDetails.secureNote
 
 import com.andryoga.safebox.BuildConfig
