+++ app/src/main/res/layout/fragment_add_new_bank_account_data_dialog.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="viewModel"
-            type="com.andryoga.safebox.ui.view.home.addNewData.bankAccount.AddNewBankAccountDataViewModel" />
+            type="com.andryoga.safebox.ui.view.home.dataDetails.bankAccount.AddNewBankAccountDataViewModel" />
     </data>
 
     <androidx.core.widget.NestedScrollView
