+++ app/src/main/res/layout/login_data_fragment.xml
@@ -7,7 +7,7 @@
     <data>
         <variable
             name="screenData"
-            type="com.andryoga.safebox.ui.view.home.addNewData.login.LoginScreenData" />
+            type="com.andryoga.safebox.ui.view.home.dataDetails.login.LoginScreenData" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
